+++ test/acceptance/app.test.js
@@ -48,6 +48,8 @@ test('GET /api/v1/sql', function(done){
     },{
         status: 400
     }, function(res) {
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         assert.deepEqual(JSON.parse(res.body), {"error":["You must indicate a sql query"]});
         done();
     });
@@ -109,6 +111,25 @@ function(done){
     });
 });
 
+// Test for https://github.com/Vizzuality/CartoDB-SQL-API/issues/85
+test("paging doesn't break x-cache-channel", 
+function(done){
+    assert.response(app, {
+        url: '/api/v1/sql?' + querystring.stringify({
+          q: 'SELECT cartodb_id*3 FROM untitle_table_4',
+          api_key: '1234',
+          rows_per_page: 1,
+          page: 2
+        }),
+        headers: {host: 'vizzuality.cartodb.com'},
+        method: 'GET'
+    },{ }, function(res) {
+        assert.equal(res.statusCode, 200, res.body);
+        assert.equal(res.headers['x-cache-channel'], 'cartodb_test_user_1_db:untitle_table_4');
+        done();
+    });
+});
+
 
 test('POST /api/v1/sql with SQL parameter on SELECT only. no database param, just id using headers', function(done){
     assert.response(app, {
@@ -122,25 +143,42 @@ test('POST /api/v1/sql with SQL parameter on SELECT only. no database param, jus
     });
 });
 
-test('GET /api/v1/sql with SQL parameter on INSERT only. oAuth not used, so public user - should fail', function(){
+test('GET /api/v1/sql with INSERT. oAuth not used, so public user - should fail', function(done){
     assert.response(app, {
         url: "/api/v1/sql?q=INSERT%20INTO%20untitle_table_4%20(id)%20VALUES%20(1)&database=cartodb_dev_user_1_db",
         method: 'GET'
     },{
-        status: 400
+    }, function(res) {
+        assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
+        assert.deepEqual(JSON.parse(res.body), 
+          // FIXME: doesn't look like this is what the test subject wants to test...
+          {"error":["relation \"untitle_table_4\" does not exist"]}
+        );
+        done();
     });
 });
 
-test('GET /api/v1/sql with SQL parameter on DROP DATABASE only. oAuth not used, so public user - should fail', function(){
+test('GET /api/v1/sql with DROP TABlE. oAuth not used, so public user - should fail', function(done){
     assert.response(app, {
         url: "/api/v1/sql?q=DROP%20TABLE%20untitle_table_4&database=cartodb_dev_user_1_db",
         method: 'GET'
     },{
-        status: 400
+    }, function(res) {
+        assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
+        assert.deepEqual(JSON.parse(res.body), 
+          // FIXME: doesn't look like this is what the test subject wants to test...
+          {"error":["table \"untitle_table_4\" does not exist"]}
+        );
+        done();
     });
 });
 
-test('GET /api/v1/sql with SQL parameter on INSERT only. header based db - should fail', function(){
+// FIXME: Duplicated test, drop 
+test('GET /api/v1/sql with INSERT. header based db - should fail', function(){
     assert.response(app, {
         url: "/api/v1/sql?q=INSERT%20INTO%20untitle_table_4%20(id)%20VALUES%20(1)",
         headers: {host: 'vizzuality.cartodb.com'},
@@ -302,13 +340,19 @@ test('DELETE with RETURNING returns all results', function(done){
     });
 });
 
-test('GET /api/v1/sql with SQL parameter on DROP DATABASE only.header based db - should fail', function(){
+test('GET /api/v1/sql with SQL parameter on DROP TABLE. should fail', function(done){
     assert.response(app, {
         url: "/api/v1/sql?q=DROP%20TABLE%20untitle_table_4",
         headers: {host: 'vizzuality.cartodb.com'},
         method: 'GET'
-    },{
-        status: 400
+    },{}, function(res) {
+        assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
+        assert.deepEqual(JSON.parse(res.body), 
+          {"error":["must be owner of relation untitle_table_4"]} 
+        );
+        done();
     });
 });
 
@@ -343,6 +387,8 @@ test('COPY TABLE with GET and auth', function(done){
     },{}, function(res) {
       // We expect a problem, actually
       assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+      assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+      assert.deepEqual(res.headers['content-disposition'], 'inline');
       assert.deepEqual(JSON.parse(res.body), {"error":["COPY from stdin failed: No source stream defined"]});
       done();
     });
@@ -359,6 +405,8 @@ test('COPY TABLE with GET and auth', function(done){
     },{}, function(res) {
       // We expect a problem, actually
       assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+      assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+      assert.deepEqual(res.headers['content-disposition'], 'inline');
       assert.deepEqual(JSON.parse(res.body), {"error":["must be superuser to COPY to or from a file"]});
       done();
     });
@@ -378,6 +426,23 @@ test('ALTER TABLE with GET and auth', function(done){
       // See https://github.com/Vizzuality/CartoDB-SQL-API/issues/43
       assert.equal(res.headers['x-cache-channel'], 'NONE');
       assert.equal(res.headers['cache-control'], expected_cache_control);
+      done();
+    });
+});
+
+test('multistatement insert, alter, select, begin, commit', function(done){
+    assert.response(app, {
+        url: "/api/v1/sql?" + querystring.stringify({
+          q: 'BEGIN; DELETE FROM test_table; COMMIT; BEGIN; INSERT INTO test_table(b) values (5); COMMIT; ALTER TABLE test_table ALTER b TYPE float USING b::float/2; SELECT b FROM test_table; COMMIT;',
+          api_key: 1234
+        }),
+        headers: {host: 'vizzuality.cartodb.com'},
+        method: 'GET'
+    },{}, function(res) {
+      assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
+      var parsedBody = JSON.parse(res.body);
+      assert.equal(parsedBody.total_rows, 1);
+      assert.deepEqual(parsedBody.rows[0], {b:2.5});
       done();
     });
 });
@@ -443,6 +508,8 @@ test('sends a 400 when an unsupported format is requested', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 400, res.body);
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         assert.deepEqual(JSON.parse(res.body), {"error":[ "Invalid format: unknown" ]});
         done();
     });
@@ -546,6 +613,8 @@ test('GET /api/v1/sql ensure cross domain set on errors', function(done){
         status: 400
     }, function(res){
         var cd = res.header('Access-Control-Allow-Origin');
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         assert.equal(cd, '*');
         done();
     });
@@ -558,7 +627,12 @@ test('cannot GET system tables', function(done){
         method: 'GET'
     },{
         status: 403
-    }, function() { done(); });
+    }, function(res) {
+      assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+      assert.deepEqual(res.headers['content-disposition'], 'inline');
+      // TODO: check actual error message...
+      done();
+    });
 });
 
 test('GET decent error if domain is incorrect', function(done){
@@ -569,6 +643,8 @@ test('GET decent error if domain is incorrect', function(done){
     },{
         status: 404
     }, function(res){
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         var result = JSON.parse(res.body);
         assert.equal(result.error[0],"Sorry, we can't find this CartoDB. Please check that you have entered the correct domain.");
         done();
@@ -584,6 +660,8 @@ test('GET decent error if SQL is broken', function(done){
         method: 'GET'
     },{}, function(res){
         assert.equal(res.statusCode, 400, res.statusCode + ': ' + res.body);
+        assert.deepEqual(res.headers['content-type'], 'application/json; charset=utf-8');
+        assert.deepEqual(res.headers['content-disposition'], 'inline');
         var result = JSON.parse(res.body);
         // NOTE: actual error message may be slighly different, possibly worth a regexp here
         assert.equal(result.error[0], 'syntax error at or near "and"');